imdb/requirements.txt
pandas==2.2.2
numpy==1.26.4
google-cloud-storage
google-cloud-bigquery
jupyter
